Reimplement script operands using OOP
This re-implements the entire script operands system to use object oriented programming instead of simple tables.

The `Operand` and `NumericOperand` classes should be used to create a new operand. The newly created `Operand` will register itself with the script operands system. It expects a unique ID and will throw an error if an `Operand` is being created with an existing ID. An environement can be provided when the `Operand` is created, otherwise the Lua environement in which the `Operand` is executed will be empty.

An `Operand` SHOULD redefine the `Operand:CodeReplacement(args)` method to return the generated code that should be used for the operand, using the arguments passed to the method.

The code has been updated to use the new OOP notation, and a temporary backward compatibility has been added to support calling the old `Operand.codeRemplacement(args)` in previous code.

Some basic unit tests have been added to the project. The first tests can be run by executing the global function `__TRP3ExtendedTests__operands()` and will check that all registered operands code generation does not error when given unexpected arguments.