Add first unit tests for operands system
This adds the very first unit tests for the operands system. These tests rely on the [WoWUnit](https://github.com/Jaliborc/WoWUnit) add-on to be running alongside the projet (the tests are ignored if the WoWUnit is missing).

The tests for the operands system should check that we are correctly getting the expected result from the operands' generated code.