Add an API endpoint for grabbing the current menuData in Extended Effects
This allows addons to grab the current menuData and patch it themselves before re-saving it using init. At the moment there appears to be not method to do this other than over-writing the menuData with the default+the third party addon's actions only, causing it to over-write / conflict if two addons want to add custom effect options.